Book Bank added.
So, what i learned:

Quarkus highly relies CDI which name is Quarkus Arc. We have to add it as a dependency over the Maven.

@ApplicationScoped and @RequestScoped two of the annotations of Arc that allows to create a bean. (@ApplicationScoped is a singleton.)

For access an any bean, we use @Inject annotation.

Also, JSON-B is a JSON serializer automatically as I used on Spring which name was Jackson.

Lastly, we have very familiar API: JAX-RS which allow to create REST. @GET, @POST, @Path or @PathParam...